14. In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	int fib1 = 1, fib2 = 2, fib3, count;
+	unsigned long fib1 = 1, fib2 = 2, fib3, count;
 
 	printf("%d, %d", fib1, fib2);
